adapt new stuff to this PR
