+++ p2p/ipld/write.go
@@ -2,14 +2,15 @@ package ipld
 
 import (
 	"context"
-	"crypto/sha256"
 	"errors"
 	"fmt"
+	"math"
 
 	format "github.com/ipfs/go-ipld-format"
 	"github.com/lazyledger/nmt"
 	"github.com/lazyledger/rsmt2d"
 
+	"github.com/lazyledger/lazyledger-core/p2p/ipld/wrapper"
 	"github.com/lazyledger/lazyledger-core/types"
 )
 
@@ -29,31 +30,23 @@ func PutBlock(ctx context.Context, adder format.NodeAdder, block *types.Block) e
 		return nil
 	}
 
+	// create nmt adder wrapping batch adder
+	batchAdder := NewNmtNodeAdder(ctx, format.NewBatch(ctx, adder))
+
+	// create the nmt wrapper to generate row and col commitments
+	squareSize := uint32(math.Sqrt(float64(len(shares))))
+	tree := wrapper.NewErasuredNamespacedMerkleTree(uint64(squareSize), nmt.NodeVisitor(batchAdder.Visit))
+
 	// recompute the eds
-	eds, err := rsmt2d.ComputeExtendedDataSquare(shares, rsmt2d.NewRSGF8Codec(), rsmt2d.NewDefaultTree)
+	eds, err := rsmt2d.ComputeExtendedDataSquare(shares, rsmt2d.NewRSGF8Codec(), tree.Constructor)
 	if err != nil {
 		return fmt.Errorf("failure to recompute the extended data square: %w", err)
 	}
 
-	// add namespaces to erasured shares and flatten the eds
-	leaves := types.FlattenNamespacedEDS(namespacedShares, eds)
-
-	// create nmt adder wrapping batch adder
-	batchAdder := NewNmtNodeAdder(ctx, format.NewBatch(ctx, adder))
-
-	// iterate through each set of col and row leaves
-	for _, leafSet := range leaves {
-		tree := nmt.New(sha256.New(), nmt.NodeVisitor(batchAdder.Visit))
-		for _, share := range leafSet {
-			err = tree.Push(share)
-			if err != nil {
-				return err
-			}
-		}
-
-		// compute the root in order to collect the ipld.Nodes
-		tree.Root()
-	}
+	// thanks to the batchAdder.Visit func we added to the nmt wrapper,
+	// generating the roots will start adding the data to IPFS
+	eds.RowRoots()
+	eds.ColumnRoots()
 
 	// commit the batch to ipfs
 	return batchAdder.Commit()
