+++ p2p/ipld/write_test.go
@@ -7,10 +7,18 @@ import (
 
 	"github.com/ipfs/go-ipfs/core/coreapi"
 	coremock "github.com/ipfs/go-ipfs/core/mock"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	mrand "math/rand"
+
+	abci "github.com/lazyledger/lazyledger-core/abci/types"
+	"github.com/lazyledger/lazyledger-core/crypto/tmhash"
 	"github.com/lazyledger/lazyledger-core/ipfs/plugin"
+	tmproto "github.com/lazyledger/lazyledger-core/proto/tendermint/types"
 	"github.com/lazyledger/lazyledger-core/types"
+	"github.com/lazyledger/lazyledger-core/types/consts"
+	"github.com/lazyledger/nmt"
 )
 
 func TestPutBlock(t *testing.T) {
@@ -24,7 +32,7 @@ func TestPutBlock(t *testing.T) {
 		t.Error(err)
 	}
 
-	maxOriginalSquareSize := types.MaxSquareSize / 2
+	maxOriginalSquareSize := consts.MaxSquareSize / 2
 	maxShareCount := maxOriginalSquareSize * maxOriginalSquareSize
 
 	testCases := []struct {
@@ -75,12 +83,120 @@ func TestPutBlock(t *testing.T) {
 	}
 }
 
+type preprocessingApp struct {
+	abci.BaseApplication
+}
+
+func (app *preprocessingApp) PreprocessTxs(
+	req abci.RequestPreprocessTxs) abci.ResponsePreprocessTxs {
+	time.Sleep(time.Second * 2)
+	randTxs := generateRandTxs(64, 256)
+	randMsgs := generateRandNamespacedRawData(128, nmt.DefaultNamespaceIDLen, 256)
+	randMessages := toMessageSlice(randMsgs)
+	return abci.ResponsePreprocessTxs{
+		Txs:      append(req.Txs, randTxs...),
+		Messages: &tmproto.Messages{MessagesList: randMessages},
+	}
+}
+func generateRandTxs(num int, size int) [][]byte {
+	randMsgs := generateRandNamespacedRawData(num, nmt.DefaultNamespaceIDLen, size)
+	for _, msg := range randMsgs {
+		copy(msg[:nmt.DefaultNamespaceIDLen], consts.TxNamespaceID)
+	}
+	return randMsgs
+}
+
+func toMessageSlice(msgs [][]byte) []*tmproto.Message {
+	res := make([]*tmproto.Message, len(msgs))
+	for i := 0; i < len(msgs); i++ {
+		res[i] = &tmproto.Message{NamespaceId: msgs[i][:nmt.DefaultNamespaceIDLen], Data: msgs[i][nmt.DefaultNamespaceIDLen:]}
+	}
+	return res
+}
+
+func TestDataAvailabilityHeaderRewriteBug(t *testing.T) {
+	ipfsNode, err := coremock.NewMockNode()
+	if err != nil {
+		t.Error(err)
+	}
+
+	ipfsAPI, err := coreapi.NewCoreAPI(ipfsNode)
+	if err != nil {
+		t.Error(err)
+	}
+	txs := types.Txs{}
+	l := len(txs)
+	bzs := make([][]byte, l)
+	for i := 0; i < l; i++ {
+		bzs[i] = txs[i]
+	}
+	app := &preprocessingApp{}
+
+	// See state.CreateProposalBlock to understand why we do this here:
+	processedBlockTxs := app.PreprocessTxs(abci.RequestPreprocessTxs{Txs: bzs})
+	ppt := processedBlockTxs.GetTxs()
+
+	pbmessages := processedBlockTxs.GetMessages()
+
+	lp := len(ppt)
+	processedTxs := make(types.Txs, lp)
+	if lp > 0 {
+		for i := 0; i < l; i++ {
+			processedTxs[i] = ppt[i]
+		}
+	}
+
+	messages := types.MessagesFromProto(pbmessages)
+	lastID := makeBlockIDRandom()
+	h := int64(3)
+
+	voteSet, _, vals := randVoteSet(h-1, 1, tmproto.PrecommitType, 10, 1)
+	commit, err := types.MakeCommit(lastID, h-1, 1, voteSet, vals, time.Now())
+	assert.NoError(t, err)
+	block := types.MakeBlock(1, processedTxs, nil, nil, messages, commit)
+	block.Hash()
+
+	hash1 := block.DataAvailabilityHeader.Hash()
+
+	ctx := context.TODO()
+	err = PutBlock(ctx, ipfsAPI.Dag(), block)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	block.Hash()
+	hash2 := block.DataAvailabilityHeader.Hash()
+	assert.Equal(t, hash1, hash2)
+
+}
+
 func generateRandomMsgOnlyData(msgCount int) types.Data {
 	out := make([]types.Message, msgCount)
-	for i, msg := range generateRandNamespacedRawData(msgCount, types.NamespaceSize, types.MsgShareSize-2) {
-		out[i] = types.Message{NamespaceID: msg[:types.NamespaceSize], Data: msg[types.NamespaceSize:]}
+	for i, msg := range generateRandNamespacedRawData(msgCount, consts.NamespaceSize, consts.MsgShareSize-2) {
+		out[i] = types.Message{NamespaceID: msg[:consts.NamespaceSize], Data: msg[consts.NamespaceSize:]}
 	}
 	return types.Data{
 		Messages: types.Messages{MessagesList: out},
 	}
+}
+
+func makeBlockIDRandom() types.BlockID {
+	var (
+		blockHash   = make([]byte, tmhash.Size)
+		partSetHash = make([]byte, tmhash.Size)
+	)
+	mrand.Read(blockHash)
+	mrand.Read(partSetHash)
+	return types.BlockID{blockHash, types.PartSetHeader{123, partSetHash}}
+}
+
+func randVoteSet(
+	height int64,
+	round int32,
+	signedMsgType tmproto.SignedMsgType,
+	numValidators int,
+	votingPower int64,
+) (*types.VoteSet, *types.ValidatorSet, []types.PrivValidator) {
+	valSet, privValidators := types.RandValidatorSet(numValidators, votingPower)
+	return types.NewVoteSet("test_chain_id", height, round, signedMsgType, valSet), valSet, privValidators
 }
